Moves rating tag to file
